% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Annotated.R
\name{config,Annotated-method}
\alias{config,Annotated-method}
\title{View PEP config of the object of \code{\link[pepr]{Project-class}}}
\usage{
\S4method{config}{Annotated}(object)
}
\arguments{
\item{object}{an object of \code{\link[pepr]{Project-class}}}
}
\value{
a list with the config file
}
\description{
This method can be used to view the config slot of
the \code{\link[pepr]{Project-class}}
 or \code{\link[S4Vectors]{Annotated-class}}
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_BiocProject", "project_config.yaml", package="BiocProject")
p=BiocProject(projectConfig)
config(p)

}
