% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkPifaceType}
\alias{.checkPifaceType}
\title{Validate type of the pipeline interface}
\usage{
.checkPifaceType(piface, type)
}
\arguments{
\item{piface}{pipeline interface to inspect}

\item{type}{string, type of the pipeline interface, either "sample" or "project"}
}
\value{
a logical indicating whether the pipeline interface matches the specified type
}
\description{
Validate type of the pipeline interface
}
