% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getBiocConfig}
\alias{.getBiocConfig}
\title{Get the preferred source of the bioconductor section}
\usage{
.getBiocConfig(p, projectLevel = FALSE)
}
\arguments{
\item{p}{\code{\link[pepr]{Project-class}} object}

\item{projectLevel}{logical indicating whether a only project-level pifaces 
should be considered. Otherwise, only sample-level ones are.}
}
\value{
a list with the selected config
}
\description{
Get the preferred source of the bioconductor section
}
