% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.getBiocConfigSrc}
\alias{.getBiocConfigSrc}
\title{Get the preferred source of the bioconductor section}
\usage{
.getBiocConfigSrc(p, pipelineName)
}
\arguments{
\item{p}{Project object}
}
\description{
Get the preferred source of the bioconductor section
}
