% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{.getOutputs}
\alias{.getOutputs}
\title{Get the outputs}
\usage{
.getOutputs(project)
}
\arguments{
\item{project}{an object of \code{\link[pepr]{Project-class}}}
}
\value{
named list of output path templates, like: \code{"aligned_{sample.genome}/{sample.sample_name}_sort.bam"}
}
\description{
Gets the pipeline outputs which are defined in the pipeline interface indicated in the \code{\link[pepr]{Project-class}}
}
