% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getPifaceByPipeline}
\alias{.getPifaceByPipeline}
\title{Get the pipeline interface with a pipeline}
\usage{
.getPifaceByPipeline(pipeline, pifaces)
}
\arguments{
\item{pipeline}{string, name of the pipeline}

\item{pifaces}{a list of pipeline interfaces \code{\link[pepr]{Config-class}} objects)}
}
\value{
piface \code{\link[pepr]{Config-class}} pipeline interface with the selected pipeline defined
}
\description{
Gets the pipeline interface which defines the pipeline from a list of pipeline interfaces
}
