% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.insertPEP}
\alias{.insertPEP}
\title{Insert a PEP metadata in a metadata slot of Annotated}
\usage{
.insertPEP(object, pep)
}
\arguments{
\item{object}{an object of \code{\link[S4Vectors]{Annotated-class}}}

\item{pep}{an object of class \code{\link[pepr]{Project-class}}}
}
\value{
an object of the same class as the object argument but enriched with the metadata from the pep argument
}
\description{
This function inserts the PEP (\code{\link[pepr]{Project-class}}) into the metadata slot of objects that extend the \code{\link[S4Vectors]{Annotated-class}}
}
\examples{
# If the object is of class Annotated
object = S4Vectors::List(result="test")
result = .insertPEP(object, pepr::Project())
metadata(result)

# If the object is not of class Annotated
object1 = "test"
result1 = .insertPEP(object1, pepr::Project())
metadata(result1)
}
