% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.makeReadFunPathAbs}
\alias{.makeReadFunPathAbs}
\title{Make readFunPath absolute}
\usage{
.makeReadFunPathAbs(piface, parent)
}
\arguments{
\item{piface}{\code{\link[pepr]{Config-class}}/list with a pipeline interface}

\item{p}{\code{\link[pepr]{Project-class}} object}
}
\value{
piface \code{\link[pepr]{Config-class}} pipeline interface with the readFunPath made absolute
}
\description{
Uses the absolute pipeline interface path in the config to determine the
absolute path to the readFunPath file that consists of the data processing function
}
