% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{.populateRecursively}
\alias{.populateRecursively}
\title{Recursively populate paths in results of type object}
\usage{
.populateRecursively(l, project, sampleName, projectContext = FALSE)
}
\arguments{
\item{l}{list to populate paths in}

\item{project}{\code{\link[pepr]{Project-class}} object}

\item{sampleName}{name of the sample to populate the outputs for}
}
\value{
list with populate paths
}
\description{
Recursively populate paths in results of type object
}
