% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{.populateSchemaPaths}
\alias{.populateSchemaPaths}
\title{Populate values in output schema}
\usage{
.populateSchemaPaths(
  schema,
  project,
  projectContext = FALSE,
  sampleName = NULL
)
}
\arguments{
\item{schema}{schema with value templates to populate}

\item{project}{\code{\link[pepr]{Project-class}} object}

\item{projectContext}{whether the values for path templates populating 
should be sourced from the project metadata. Otherwise metadata for 
a selected sample is used}

\item{sampleName}{name of the sample to populate the outputs for. Required 
if \code{projectContext} set to \code{FALSE}}
}
\value{
a possibly nested list of length equal to the number of results defined in 
the schema with populated outputs
}
\description{
Populates schema values of type path and thumbnail path in the provided 
output schema for the selected sample or project
}
