% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{.pyToR}
\alias{.pyToR}
\title{Switches from python to R list accession syntax}
\usage{
.pyToR(str)
}
\arguments{
\item{str}{String to recode}
}
\value{
string with the recoded accession syntax
}
\description{
Python uses a dot to access attributes, while R uses \code{$}; this function
converts the python style into R so that we can use R code to populate
variables with R lists. From this: '\code{sample.name}' to this: '\code{sample$name}'
}
