% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.unionList}
\alias{.unionList}
\title{Update list with another list}
\usage{
.unionList(list1, list2, combine = FALSE)
}
\arguments{
\item{list1}{a list to be updated}

\item{list2}{a list to update with}

\item{combine}{a logical indicating whether the elements of the second list 
should replace (\code{FALSE}, default) or append to (\code{TRUE}) the 
first one.}
}
\value{
an updated list
}
\description{
This function performs a union of two lists and updates the elements of the 
first one if are found in the other one.
}
\details{
Both elements have to be lists. If some elements are not named, they are 
preserved but the order might be lost.
}
\examples{
list1=list(a=1,b=2)
list2=list(a=1,b=1,c=3)
.unionList(list1,list2)

}
