% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{gatherPipelineInterfaces,Project-method}
\alias{gatherPipelineInterfaces,Project-method}
\title{Collect all pipeline interfaces}
\usage{
\S4method{gatherPipelineInterfaces}{Project}(project, projectLevel = FALSE)
}
\arguments{
\item{project}{\code{\link[pepr]{Project-class}} object}

\item{projectLevel}{logical indicating whether a only project-level pifaces 
should be considered. Otherwise, only sample-level ones are.}
}
\value{
a list of pipeline interface file paths.
}
\description{
Collects all relevant pipeline interfaces for this \code{\link[pepr]{Project-class}}
}
\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_piface",
"project_config.yaml",
package = "BiocProject")
p = Project(file = projectConfig)
gatherPipelineInterfaces(p)
gatherPipelineInterfaces(p, TRUE)
}
