% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getMetadata,BiocProject-method}
\alias{getMetadata,BiocProject-method}
\title{Get metadata from the Project object (PEP)}
\usage{
\S4method{getMetadata}{BiocProject}(.Object)
}
\arguments{
\item{.Object}{An object of \linkS4class{BiocProject} class}
}
\value{
.Object An object of \linkS4class{BiocProject} class. The metadata slot of \linkS4class{BiocProject} is enriched with data from \linkS4class{Project} object as an additional \code{PEP config file} field list in metadata slot.
}
\description{
This method copies the \href{https://pepkit.github.io/docs/project_config/}{PEP config file} data from \linkS4class{Project} object to a list in the metadata slot of the \code{BiocProject} object.
}
