% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{getPiface}
\alias{getPiface}
\title{Returns pipeline interface object
TODO: handle multiple pipeline interfaces
TODO: Load this at construction time.}
\usage{
getPiface(project)
}
\description{
Returns pipeline interface object
TODO: handle multiple pipeline interfaces
TODO: Load this at construction time.
}
