% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{getPipelineInterfaces}
\alias{getPipelineInterfaces}
\alias{getPipelineInterfaces,Project-method}
\title{Get the pipeline intraface(s)}
\usage{
getPipelineInterfaces(.Object)

\S4method{getPipelineInterfaces}{Project}(.Object)
}
\arguments{
\item{.Object}{an object of \code{\link{Project-class}}}
}
\value{
a list of objects of \code{\link{Config-class}} 
(list-like representation of the YAML file(s)). Number of elements in the 
list reflects the number of pipeline interfaces defined by the 
\code{\link{Project-class}} object.
}
\description{
Extracts the pipeline interface(s) as a list 
of \code{\link{Config-class}} objects
}
\section{Methods (by class)}{
\itemize{
\item \code{Project}: extracts pipeline interfaces defined in a \code{\link{Project-class}} object
}}

\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_piface",
"project_config.yaml",
package = "BiocProject")
p = Project(file = projectConfig)
getPipelineInterfaces(p)
}
