% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\docType{methods}
\name{getPipelines}
\alias{getPipelines}
\alias{getPipelines,Config-method}
\title{Get pipelines defined within a pipeline interface}
\usage{
getPipelines(.Object)

\S4method{getPipelines}{Config}(.Object)
}
\arguments{
\item{.Object}{a pipeline interface, an object of \code{\link[pepr]{Config-class}}}
}
\value{
a list of pipelines, objets of \code{\link[pepr]{Config-class}}
}
\description{
Get pipelines defined within a pipeline interface
}
\section{Methods (by class)}{
\itemize{
\item \code{Config}: extracts pipelines from a pipeline interface
}}

\examples{
# add example
}
