% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Annotated.R
\docType{methods}
\name{getProject,Annotated-method}
\alias{getProject,Annotated-method}
\title{Extract the object of \code{\link[pepr]{Project-class}} from the \code{\link[S4Vectors]{Annotated-class}}}
\usage{
\S4method{getProject}{Annotated}(.Object)
}
\arguments{
\item{object}{an object of \code{\link[S4Vectors]{Annotated-class}}}
}
\value{
an object of \code{\link[pepr]{Project-class}}
}
\description{
This method can be used to extract the project metadata from objects of \code{\link[S4Vectors]{Annotated-class}}
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_BiocProject", "project_config.yaml", package="BiocProject")
p=BiocProject(projectConfig)
getProject(p)

}
