% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{getProjectOutputs}
\alias{getProjectOutputs}
\alias{getProjectOutputs,Project-method}
\title{Populates and returns outputs for a given \code{\link[pepr]{Project-class}}}
\usage{
getProjectOutputs(project)

\S4method{getProjectOutputs}{Project}(project)
}
\arguments{
\item{project}{\code{\link[pepr]{Project-class}} object}
}
\value{
a list of output file paths. The order of the first level of the
list corresponds to the order of the pipeline interface files, second level 
is a named list of file paths populated 
by the \code{\link[pepr]{Project-class}}
}
\description{
Returns the project level pipeline outputs which are defined in the output 
schema indicated by the pipeline interface indicated in the 
\code{\link[pepr]{Project-class}}
}
\section{Methods (by class)}{
\itemize{
\item \code{Project}: Populates and returns output files for 
a given \code{\link[pepr]{Project-class}}
}}

\examples{
projectConfig = system.file('extdata',
'example_peps-master',
'example_piface',
'project_config.yaml',
package = 'BiocProject')
p = Project(file = projectConfig)
getProjectOutputs(p)
}
