% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Annotated.R
\docType{methods}
\name{is,Annotated-method}
\alias{is,Annotated-method}
\title{Is an Object from a Class?}
\usage{
\S4method{is}{Annotated}(object, class2)
}
\description{
Functions to test inheritance relationships between an object and a class or between two classes. It uses the generic is function but overrides its behavior for obejcts of class \code{\link[S4Vectors]{Annotated-class}} when testing for inheritance from \code{\link[pepr]{Project-class}} class.
}
\details{
see the \code{\link[methods]{is}} for more details
}
