% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{loadPEP}
\alias{loadPEP}
\title{Load the PEP into the \code{BiocProject}}
\usage{
loadPEP(.Object, file, subproject)
}
\arguments{
\item{.Object}{object of class \code{\link{BiocProject}}}

\item{file}{a character vecotr with a path to the config file}

\item{subproject}{a character vector with a name of the subproject to be activated}
}
\value{
object of class \code{\link{BiocProject}} with the metadata slot populated with the Project object
}
\description{
This function loads the PEP into the metadata slot of the \code{\link{BiocProject}} object using the \code{\link[pepr]{Project}} constructor of the pepr package
}
\details{
This funciton is used in the \code{\link{BiocProject}} initialize method and should not be used outside
}
\seealso{
\url{https://pepkit.github.io/}
}
