% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{metadata,BiocProject-method}
\alias{metadata,BiocProject-method}
\title{Display the metadata}
\usage{
\S4method{metadata}{BiocProject}(x)
}
\arguments{
\item{x}{An object of \code{BiocProject} class}
}
\description{
This method displays the metadata associated with hthe \code{BiocProject} object.
}
