% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{outputsByPipelines}
\alias{outputsByPipelines}
\title{Populates and returns output files for a given pipeline}
\usage{
outputsByPipelines(project, pipelineName = NULL)
}
\arguments{
\item{project}{\code{\link[pepr]{Project-class}} object}

\item{pipelineNames}{pipeline name to return the outputs for}
}
\value{
a list of output file paths for the requested pipeline
}
\description{
Returns the pipeline outputs which are defined in the pipeline interface 
indicated in the \code{\link[pepr]{Project-class}}
}
\details{
Only the samples that have a matching protocol attribute will be used to
populate the output paths.
}
\note{
If there are multiple pipeline interfaces that defined the sampe 
pipeline (name), the output files of first one will be returned
}
\examples{
#add examples
}
