% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\docType{methods}
\name{outputsByProtocols}
\alias{outputsByProtocols}
\alias{outputsByProtocols,Project-method}
\title{Populates and returns output files for a given protocol}
\usage{
outputsByProtocols(project, ...)

\S4method{outputsByProtocols}{Project}(project, protocolNames = NULL)
}
\arguments{
\item{project}{\code{\link[pepr]{Project-class}} object}

\item{...}{other arguments passed to methods}

\item{protocolNames}{char vector of protocol names to match the pipelines 
and return their outputs}
}
\value{
a list of output file paths. The order of the first level of the 
list corresponds to the order of the pipeline interface files, second level 
order (named) reflects the pipelines within the files, the last level is a 
named list of file paths populated by the samples
}
\description{
Returns the pipeline outputs which are defined in the pipeline interface 
indicated in the \code{\link[pepr]{Project-class}}
}
\section{Methods (by class)}{
\itemize{
\item \code{Project}: extracts pipeline outputs for a given protocol or set of protocols
}}

\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_piface",
"project_config.yaml",
package = "BiocProject")
p = Project(file = projectConfig)
outputsByProtocols(p, "PROTO2")
}
