% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{outputsBySample,Project-method}
\alias{outputsBySample,Project-method}
\title{and return their outputs}
\usage{
\S4method{outputsBySample}{Project}(project, sampleNames = NULL)
}
\description{
and return their outputs
}
