% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{outputsBySample}
\alias{outputsBySample}
\title{Populates and returns output files for a given sample}
\usage{
outputsBySample(project, ...)
}
\arguments{
\item{project}{\code{\link[pepr]{Project-class}} object}

\item{...}{other arguments passed to methods}
}
\value{
a list of output file paths
}
\description{
Returns the pipeline outputs which are defined in the pipeline interface
indicated in the \code{\link[pepr]{Project-class}}
}
\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_piface",
"project_config.yaml",
package = "BiocProject")
p = Project(file = projectConfig)
outputsBySample(p, "PROTO2")
}
