% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{pipelineInterfacesBySample,Project-method}
\alias{pipelineInterfacesBySample,Project-method}
\title{Get pipeline interfaces by sample}
\usage{
\S4method{pipelineInterfacesBySample}{Project}(project)
}
\arguments{
\item{project}{\code{\link[pepr]{Project-class}} object}

\item{...}{other arguments passed to methods}
}
\value{
a list of pipeline interface file paths keyed by sample names
}
\description{
Collects all relevant pipeline interfaces for this 
\code{\link[pepr]{Project-class}} and provides a sample to interfaces mapping
}
\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_piface",
"project_config.yaml",
package = "BiocProject")
p = Project(file = projectConfig)
pipelineInterfacesBySample(p)
}
