% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{pytor}
\alias{pytor}
\title{Switches from python to R list accession syntax}
\usage{
pytor(str)
}
\arguments{
\item{str}{String to recode}
}
\description{
Switches from python to R list accession syntax
}
\examples{
pytor("{sample.name}...{project.source.name}")
}
