% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readSchema}
\alias{readSchema}
\title{Read a YAML-formatted schema}
\usage{
readSchema(path, parent = NULL)
}
\arguments{
\item{path}{path to a local schema or URL pointing to a remote one}

\item{parent}{a path to parent folder to use}
}
\value{
list read schema
}
\description{
Remote or local schemas are supported
}
\examples{
readSchema('https://schema.databio.org/pep/2.0.0.yaml')
}
