% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Annotated.R
\docType{methods}
\name{samples,Annotated-method}
\alias{samples,Annotated-method}
\title{View samples in the objects of \code{\link[pepr]{Project-class}}}
\usage{
\S4method{samples}{Annotated}(object)
}
\arguments{
\item{object}{an object of \code{\link[pepr]{Project-class}}}
}
\value{
a data.table with the with metadata about samples
}
\description{
This method can be used to view the samples slot
of the \code{\link[pepr]{Project-class}} 
or \code{\link[S4Vectors]{Annotated-class}}
}
\examples{
projectConfig = system.file("extdata", "example_peps-master",
"example_BiocProject", "project_config.yaml", package="BiocProject")
p=BiocProject(projectConfig)
samples(p)

}
