% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{samplesByProtocol}
\alias{samplesByProtocol}
\title{Get samples that match the protocol}
\usage{
samplesByProtocol(s, protocolName, caseSensitive = FALSE)
}
\arguments{
\item{s}{samples, e.g. output of samples(project)}

\item{protocolName, }{string name of the protocol}

\item{caseSensitive, }{logical indicatinh whether the protocol match should be case sensitive}
}
\value{
a subset of samples that match the protocol
}
\description{
Get samples that match the protocol
}
\examples{
#add examples
}
