% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{toProject}
\alias{toProject}
\title{Get \code{\link[pepr]{Project-class}} object}
\usage{
toProject(.Object)
}
\arguments{
\item{.Object}{An object of \code{\link{BiocProject-class}}}
}
\value{
an object of \code{\link[pepr]{Project-class}} object
}
\description{
This method coerces the \code{\link{BiocProject-class}}
 to \code{\link{Project-class}}
}
\examples{
ProjectConfig = system.file(
"extdata",
"example_peps-master",
"example_BiocProject",
"project_config.yaml",
package = "BiocProject"
)

bp = BiocProject(file=ProjectConfig)
toProject(bp)

}
