% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{BiocProject-class}
\alias{BiocProject-class}
\title{Portable Encapsulated Project (PEP) class for biological applications}
\description{
This class provides a link between PEP and biological data structures
}
\section{Slots}{

\describe{
\item{\code{metadata}}{see see \link[S4Vectors]{Annotated} class for details. Meant to store an object of class \link[pepr]{Project}}
}}

