% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{BiocProject-class}
\alias{BiocProject-class}
\title{A class representing a Portable Encapsulated Project and Summarized Experiment objects interface}
\description{
Provides a representation and functions to access project
configuration and sample annotation values for a PEP as well as functions concerning experimental results organization provided by the \linkS4class{RangedSummarizedExperiment} class.
Additionally, this class privides an interfece that connects them.
}
\details{
This class inherits from classes \linkS4class{Project} and \linkS4class{RangedSummarizedExperiment}
}
\section{Slots}{

\describe{
\item{\code{ProjectSlots}}{see \link[pepr]{Project} class for details}

\item{\code{SummarizedExperimentSlots}}{see \link[SummarizedExperiment]{SummarizedExperiment} class docs for details}
}}

