% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{BiocProject}
\alias{BiocProject}
\title{A class representing a Portable Encapsulated Project and Summarized Experiment objects interface
This is a helper that creates the BiocProject with empty Project and SummarizedExperiment objects included}
\usage{
BiocProject(file = character(), ...)
}
\arguments{
\item{file}{a string with a path to the config file as in \linkS4class{Project}}
}
\description{
A class representing a Portable Encapsulated Project and Summarized Experiment objects interface
This is a helper that creates the BiocProject with empty Project and SummarizedExperiment objects included
}
