% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{BiocProject}
\alias{BiocProject}
\title{Portable Encapsulated Project (PEP) class for biological applications}
\usage{
BiocProject(file = character(), subproject = character(), ...)
}
\arguments{
\item{file}{a character vecotr with a path to the config file}

\item{subproject}{a character vector with a name of the subproject to be activated}
}
\value{
an object of \code{\link{BiocProject}} class
}
\description{
This is a helper that creates the \code{\link{BiocProject}} object
}
\seealso{
\url{https://pepkit.github.io/}
}
