% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getBiocConfig}
\alias{.getBiocConfig}
\title{Get the preferred source of the bioconductor section}
\usage{
.getBiocConfig(p, pipelineName)
}
\arguments{
\item{p}{\code{\link[pepr]{Project-class}} object}

\item{pipelineName}{name of the pipeline within pipeline interface where 
the 'bioconductor' section should be searched for.}
}
\value{
a list with the selected config
}
\description{
Get the preferred source of the bioconductor section
}
