% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{.getOutputs}
\alias{.getOutputs}
\title{Get outputs from pipeline defined in an output schema}
\usage{
.getOutputs(pipeline, projectContext = FALSE, parent)
}
\arguments{
\item{pipeline}{an object of \code{\link[pepr]{Config-class}}}
}
\value{
named list of output path templates, 
like: \code{"aligned_{sample.genome}/{sample.sample_name}_sort.bam"}
}
\description{
Extracts the output file templates defined for a given pipeline
}
