% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.mkPathsAbs}
\alias{.mkPathsAbs}
\title{Make all paths absolute by parent dir}
\usage{
.mkPathsAbs(paths, parent)
}
\arguments{
\item{paths}{path to make absolute}

\item{parent}{parent directory}
}
\value{
string, a path amde absolute
}
\description{
Make all paths absolute by parent dir
}
