% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{.populateString}
\alias{.populateString}
\title{Populates a variable-encoded string with sample/project variables}
\usage{
.populateString(string, project)
}
\arguments{
\item{string}{Variable-encoded string to populate}

\item{\code{\link[pepr]{Project-class}}}{object with values to draw from}

\item{protocolName}{string, name of the protocol to select the samples}
}
\value{
populated string
}
\description{
Given a string and a project this function will go through samples and populate
the variables. Used to return real files for each sample from an output variable
in the pipeline interface
}
\examples{
configFile = system.file(
  "extdata",
  "example_peps-master",
  "example_BiocProject",
  "project_config.yaml",
  package = "BiocProject"
)
bp = BiocProject(file=configFile)
.populateString("{sample.sample_name}/{sample.file_path}", bp)
}
