% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.updateList}
\alias{.updateList}
\title{Update list with another list}
\usage{
.updateList(list1, list2)
}
\arguments{
\item{list1}{a list to be updated}

\item{list2}{a list to update with}
}
\value{
an updated list
}
\description{
This function performs a union of two lists and updates the elements of the first one if are found in the other one
}
