% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{gatherPipelineInterfaces}
\alias{gatherPipelineInterfaces}
\alias{gatherPipelineInterfaces,Project-method}
\alias{.gatherSamplePipelineInterfaces,Project-method}
\title{Collect all pipeline interfaces}
\usage{
gatherPipelineInterfaces(project, ...)

\S4method{gatherPipelineInterfaces}{Project}(project, projectLevel = FALSE)

\S4method{.gatherSamplePipelineInterfaces}{Project}(project)
}
\arguments{
\item{project}{\code{\link[pepr]{Project-class}} object}

\item{...}{other arguments}

\item{projectLevel}{logical indicating whether a only project-level pifaces 
should be considered. Otherwise, only sample-level ones are.}
}
\value{
a list of pipeline interface file paths.
}
\description{
Collects all relevant pipeline interfaces for this \code{\link[pepr]{Project-class}}
}
\section{Methods (by class)}{
\itemize{
\item \code{Project}: Collect all pipeline interfaces

\item \code{Project}: extracts pipeline outputs for a given pipeline
}}

\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_piface",
"project_config.yaml",
package = "BiocProject")
p = Project(file = projectConfig)
gatherPipelineInterfaces(p)
gatherPipelineInterfaces(p, TRUE)
}
