% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getColData,BiocProject-method}
\alias{getColData,BiocProject-method}
\title{Get colData from the Project object (PEP)}
\usage{
\S4method{getColData}{BiocProject}(.Object)
}
\arguments{
\item{.Object}{An object of \linkS4class{BiocProject} class}
}
\value{
.Object An object of \linkS4class{BiocProject} class. The colData slot is derived from samples attribute of \linkS4class{Project}
}
\description{
This method copies info about samples from \linkS4class{Project} object to \code{colData} slot on the \code{BiocProject} object.
}
