% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getData}
\alias{getData}
\title{Extract data from \code{\link[pepr]{Project-class}} objects}
\usage{
getData(.Object)
}
\arguments{
\item{.Object}{an object of \code{\link{BiocProject-class}}}
}
\value{
a list with the data elements
}
\description{
This method extracts the data from \code{\link[pepr]{Project-class}} objects
}
\examples{
ProjectConfig = system.file(
"extdata",
"example_peps-master",
"example_BiocProject",
"project_config.yaml",
package = "BiocProject"
)

bp = BiocProject(file=ProjectConfig)
getData(bp)

}
