% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getMetadata,BiocProject-method}
\alias{getMetadata,BiocProject-method}
\title{Get metadata from the Project object (PEP)}
\usage{
\S4method{getMetadata}{BiocProject}(.Object)
}
\arguments{
\item{.Object}{An object of BiocProject class}
}
\value{
.Object An object of BiocProject class. The metadata slot of \code{BiocProject} is enriched with data from \linkS4class{Project} object as an additional \code{PEP config file} field in the MIAME object.
}
\description{
This method copies metadata from \linkS4class{Project} object to \linkS4class{MIAME} in the \code{BiocProject} object.
}
