% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{getOutFiles}
\alias{getOutFiles}
\title{Populates and returns output files}
\usage{
getOutFiles(project)
}
\arguments{
\item{project}{\code{\link[pepr]{Project-class}} object}
}
\value{
a list of output file paths
}
\description{
Returns the pipeline outputs which are defined in the pipeline interface indicated in the \code{\link[pepr]{Project-class}}
}
\examples{
#add examples
}
