% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{getOutFiles}
\alias{getOutFiles}
\title{Populates and returns output files}
\usage{
getOutFiles(project)
}
\arguments{
\item{project}{pepr::Project object}
}
\description{
Populates and returns output files
}
