% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{getOutFiles}
\alias{getOutFiles}
\title{Populates and returns output files}
\usage{
getOutFiles(project, protocolNames = NULL)
}
\arguments{
\item{project}{\code{\link[pepr]{Project-class}} object}

\item{protocolNames}{char vector of protocol names to match the pipelines 
and return their outputs}

\item{pipelineNames}{char vector of pipeline names to return the outputs for}
}
\value{
a list of output file paths. The order of the first level of the 
list corresponds to the order of the pipeline interface files, second level 
order (named) reflects the pipelines within the files.
}
\description{
Returns the pipeline outputs which are defined in the pipeline interface 
indicated in the \code{\link[pepr]{Project-class}}
}
\examples{
#add examples
}
