% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{getPipelines}
\alias{getPipelines}
\title{Get pipelines by protocol name}
\usage{
getPipelines(.Object, protocolName = NULL)
}
\arguments{
\item{.Object}{a pipeline interface}

\item{protocolName}{a string or vector of strings indicating the protocols 
for which the pipelines should be returned}
}
\value{
a list of pipelines, objets of \code{\link[pepr]{Config-class}}
}
\description{
Gets the pipelines defined within the pipeline interface
}
\details{
To get all the pipelines (default) do not specify any \code{protocolName}
}
\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_piface",
"project_config.yaml",
package = "BiocProject")
p = Project(file = projectConfig)
pifaces = getPipelineInterfaces(p)
getPipelines(pifaces[[1]])
}
