% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\name{getProjectOutputs}
\alias{getProjectOutputs}
\title{Populates and returns output files for a given  \code{\link[pepr]{Project-class}}}
\usage{
getProjectOutputs(project)
}
\arguments{
\item{project}{\code{\link[pepr]{Project-class}} object}

\item{sampleNames}{names of the samples}
}
\value{
a list of output file paths. The order of the first level of the
list corresponds to the order of the pipeline interface files, second level 
is a named list of file paths populated by the \code{\link[pepr]{Project-class}}
}
\description{
Returns the pipeline outputs which are defined in the pipeline interface
indicated in the \code{\link[pepr]{Project-class}}
}
\examples{
projectConfig = system.file("extdata",
"example_peps-master",
"example_piface",
"project_config.yaml",
package = "BiocProject")
p = Project(file = projectConfig)
getProjectOutputs(p)
getProjectOutputs(p, "sample1")
}
