% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\docType{methods}
\name{getProtocolMappings}
\alias{getProtocolMappings}
\alias{getProtocolMappings,Config-method}
\title{Get protocol mappings defined within a pipeline interface}
\usage{
getProtocolMappings(.Object)

\S4method{getProtocolMappings}{Config}(.Object)
}
\arguments{
\item{.Object}{a pipeline interface, an object of \code{\link[pepr]{Config-class}}}
}
\value{
a named list of pipelines, where the names of the elements are the 
protocols and the elements names of the pipelines
}
\description{
Get protocol mappings defined within a pipeline interface
}
\section{Methods (by class)}{
\itemize{
\item \code{Config}: extracts protocol mappings from a pipeline interface
}}

\examples{
# add example
}
