% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Annotated.R
\docType{methods}
\name{is,Annotated-method}
\alias{is,Annotated-method}
\title{Is an Object from a Class?}
\usage{
\S4method{is}{Annotated}(object, class2)
}
\arguments{
\item{object}{the object to be tested}

\item{class2}{the class name to test the object against}
}
\value{
a logical
}
\description{
Functions to test inheritance relationships between an object and a class or between two classes. It uses the generic is function but overrides its behavior for obejcts of class \code{\link[S4Vectors]{Annotated-class}} when testing for inheritance from \code{\link[pepr]{Project-class}} class.
}
\details{
see the \code{\link[methods]{is}} for more details
}
\examples{
object = S4Vectors::List(test="test")
is(object,"Annotated")

}
