% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_interface.R
\docType{methods}
\name{outputsByPipeline}
\alias{outputsByPipeline}
\alias{outputsByPipeline,Project-method}
\title{Populates and returns output files for a given pipeline}
\usage{
outputsByPipeline(project, ...)

\S4method{outputsByPipeline}{Project}(project, pipelineName = NULL)
}
\arguments{
\item{project}{\code{\link[pepr]{Project-class}} object}

\item{pipelineName}{pipeline name to return the outputs for}
}
\value{
a named list of output file paths for the requested pipeline
}
\description{
Returns the pipeline outputs which are defined in the pipeline interface 
indicated in the \code{\link[pepr]{Project-class}}
}
\details{
Only the samples that have a matching protocol attribute will be used to
populate the output paths.
}
\section{Methods (by class)}{
\itemize{
\item \code{Project}: extracts pipeline outputs for a given pipeline
}}

\note{
If there are multiple pipeline interfaces that defined the same 
pipeline (name), the output files of first one will be returned. 
\cr \cr If no pipeline name provided, output for all pipelines with unique 
names are returned
}
\examples{
#add examples
}
